server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
opencv-python
pillow
